# THURSDAY NIGHT/definitions/swagger.yaml
swagger: '2.0'
info:
  description: Service calls for various operations
  version: 1.0.0
  title: Application API
host: localhost
basePath: /api/v1
schemes:  
  - https
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: query
    description: 
     For accessing the API a valid JWT token must be passed in all the queries in
     the query string.  
     A valid JWT token is generated by the API and returned as answer of a call
     to the route /login giving a valid user & password.
consumes:
  - application/json      

paths:
  /user/userLogin:
    post:
      consumes:
        - application/x-www-form-urlencoded
      tags:
        - User
      summary: To check whether user exists in userpool or not and returns its data.
      description: to check whether user exists in userpool or not and returns its data
      operationId: userLoginPOST
      parameters:
        - name: username
          in: formData
          description: username given during login
          required: true
          type: string
        - name: password
          in: formData
          description: password given during login
          required: true
          type: string
      responses:
        '200':
          description: Returns user id, type and other information about the user according to its type.
          schema:
            $ref: '#/definitions/userLoginResponse'
        '401':
          description: Returns error message in case of unauthorized login
          schema:
            $ref: '#/definitions/genericErrorResponse'
      x-swagger-router-controller: userService
  /user/locked:
    get:
    
      produces:
        - application/x-www-form-urlencoded
      tags:
        - admin
      summary: To check whether user exists in userpool or not and returns its data.
      description: to check whether user exists in userpool or not and returns its data
      operationId: userLocked
     
      responses:
        '200':
          description: Returns user id, type and other information about the user according to its type.
          schema:
            $ref: '#/definitions/chequeApplyResponse'
        '401':
          description: Returns error message in case of unauthorized login
          schema:
            $ref: '#/definitions/genericErrorResponse'
      x-swagger-router-controller: userLockService
  
definitions:
  userLoginResponse:
    properties:
      success:
        type: boolean
        description: "true for success. Other as per message!"
      message:
        type: string
        description: 'Will be "success" or the error message details, if applicable'
      token:
        type: string
        
  chequeApplyResponse:
    properties:
      success:
        type: boolean
        description: "true for success. Other as per message!"
      message:
        type: string
        description: 'Will be "success" or the error message details, if applicable'
      data:
        type: object
        properties:
         
          chequeNo:
            type: string
            description: "chequeNumber"
          Amount:
            type: string
            description: "Amount of application"
          account:
            type: string
            description: "account of application"
          AccountNumber:
            type: string
            description: "AccountNumber of application"
          username:
            type: string
            description: "username of application"
         
  genericErrorResponse:
    properties:
      success:
        type: boolean
        description: "true for success, false for error."
      message:
        type: string
        description: "Error message"